+++ flutter_inappwebview_macos/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -2648,6 +2648,8 @@ class MacOSInAppWebViewController extends PlatformInAppWebViewController
 
   @override
   Future<void> setJavaScriptBridgeName(String bridgeName) async {
+    assert(RegExp(r'^[a-zA-Z_]\w*$').hasMatch(bridgeName),
+      'bridgeName must be a non-empty string with only alphanumeric and underscore characters. It can\'t start with a number.');
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('bridgeName', () => bridgeName);
     await _staticChannel.invokeMethod('setJavaScriptBridgeName', args);