+++ flutter_inappwebview_windows/windows/in_app_webview/in_app_webview.h
@@ -16,6 +16,7 @@
 #include "../types/ssl_certificate.h"
 #include "../types/url_request.h"
 #include "../types/web_history.h"
+#include "../utils/util.h"
 #include "../webview_environment/webview_environment.h"
 #include "in_app_webview_settings.h"
 #include "user_content_controller.h"
@@ -192,6 +193,7 @@ namespace flutter_inappwebview_plugin
     POINT lastCursorPos_ = { 0, 0 };
     VirtualKeyState virtualKeys_;
 
+    const std::string expectedBridgeSecret = get_uuid();
     std::map<UINT64, std::shared_ptr<NavigationAction>> navigationActions_ = {};
     std::shared_ptr<NavigationAction> lastNavigationAction_;
     bool isLoading_ = false;
@@ -200,6 +202,7 @@ namespace flutter_inappwebview_plugin
 
     void registerEventHandlers();
     void registerSurfaceEventHandlers();
+    HRESULT onCallJsHandler(const bool& isMainFrame, ICoreWebView2WebMessageReceivedEventArgs* args);
   };
 }
 #endif //FLUTTER_INAPPWEBVIEW_PLUGIN_IN_APP_WEBVIEW_H_