+++ flutter_inappwebview_windows/windows/in_app_webview/user_content_controller.cpp
@@ -343,7 +343,7 @@ namespace flutter_inappwebview_plugin
 
     std::string source = userScript->source;
     if (userScript->injectionTime == UserScriptInjectionTime::atDocumentEnd) {
-      source = replace_all_copy(USER_SCRIPTS_AT_DOCUMENT_END_WRAPPER_JS_SOURCE, VAR_PLACEHOLDER_VALUE, source);
+      source = replace_all_copy(UserContentController::USER_SCRIPTS_AT_DOCUMENT_END_WRAPPER_JS_SOURCE(), VAR_PLACEHOLDER_VALUE, source);
       std::ostringstream address;
       address << std::addressof(userScript);
       replace_all(source, VAR_PLACEHOLDER_MEMORY_ADDRESS_VALUE, address.str());