+++ flutter_inappwebview_windows/windows/in_app_webview/webview_channel_delegate.cpp
@@ -377,7 +377,7 @@ namespace flutter_inappwebview_plugin
     channel->InvokeMethod("onUpdateVisitedHistory", std::move(arguments));
   }
 
-  void WebViewChannelDelegate::onCallJsHandler(const std::string& handlerName, const std::string& args, std::unique_ptr<CallJsHandlerCallback> callback) const
+  void WebViewChannelDelegate::onCallJsHandler(const std::string& handlerName, const std::unique_ptr<JavaScriptHandlerFunctionData> data, std::unique_ptr<CallJsHandlerCallback> callback) const
   {
     if (!channel) {
       callback->defaultBehaviour(std::nullopt);
@@ -386,7 +386,7 @@ namespace flutter_inappwebview_plugin
 
     auto arguments = std::make_unique<flutter::EncodableValue>(flutter::EncodableMap{
       {"handlerName", handlerName},
-      {"args", args}
+      {"data", data->toEncodableMap()}
       });
     channel->InvokeMethod("onCallJsHandler", std::move(arguments), std::move(callback));
   }