+++ flutter_inappwebview_windows/windows/types/plugin_script.cpp
@@ -6,13 +6,13 @@ namespace flutter_inappwebview_plugin
     const std::optional<std::string>& groupName,
     const std::string& source,
     const UserScriptInjectionTime& injectionTime,
-    const std::vector<std::string>& allowedOriginRules,
+    const bool& forMainFrameOnly,
+    const std::optional<std::vector<std::string>>& allowedOriginRules,
     std::shared_ptr<ContentWorld> contentWorld,
     const bool& requiredInAllContentWorlds
-  ) : UserScript(groupName, source, injectionTime, allowedOriginRules, std::move(contentWorld)),
+  ) : UserScript(groupName, source, injectionTime, forMainFrameOnly, allowedOriginRules, std::move(contentWorld)),
     requiredInAllContentWorlds_(requiredInAllContentWorlds)
   {}
-
 
   std::shared_ptr<PluginScript> PluginScript::copyAndSet(const std::shared_ptr<ContentWorld> cw) const
   {
@@ -20,6 +20,7 @@ namespace flutter_inappwebview_plugin
       this->groupName,
       this->source,
       this->injectionTime,
+      this->forMainFrameOnly,
       this->allowedOriginRules,
       cw,
       this->requiredInAllContentWorlds_