+++ flutter_inappwebview_platform_interface/CHANGELOG.md
@@ -3,7 +3,7 @@
 - Updated static `fromMap` implementation for some classes
 - Updated `kJavaScriptHandlerForbiddenNames` list
 - Added `PlatformInAppLocalhostServer.onData` parameter to set a custom on data server callback
-- Added `javaScriptHandlerOriginAllowList`, `pluginScriptsOriginAllowList`, `pluginScriptsForMainFrameOnly` InAppWebViewSettings parameters
+- Added `javaScriptHandlerOriginAllowList`, `pluginScriptsOriginAllowList`, `pluginScriptsForMainFrameOnly`, `scrollMultiplier` InAppWebViewSettings parameters
 - Added `setJavaScriptBridgeName`, `getJavaScriptBridgeName` static WebView controller methods
 - Added `JavaScriptHandlerFunctionData` type
 - Deprecated `JavaScriptHandlerCallback` type in favor of `JavaScriptHandlerFunction` type